refactor(frontend): share auth fetch helpers and guard stale workstation loads
