+++ Frontend/src/lib/auth-fetch.ts
@@ -1,0 +1,43 @@
+export function mergeSignals(signals: Array<AbortSignal | null | undefined>): AbortSignal | undefined {
+  const activeSignals = signals.filter(Boolean) as AbortSignal[];
+  if (activeSignals.length === 0) {
+    return undefined;
+  }
+  if (activeSignals.length === 1) {
+    return activeSignals[0];
+  }
+
+  const controller = new AbortController();
+  const abort = () => controller.abort();
+  for (const signal of activeSignals) {
+    if (signal.aborted) {
+      controller.abort();
+      break;
+    }
+    signal.addEventListener("abort", abort, { once: true });
+  }
+  return controller.signal;
+}
+
+export async function authFetch(
+  getToken: () => Promise<string | null>,
+  input: string,
+  init: RequestInit = {},
+  timeoutMs = 10000,
+): Promise<Response> {
+  const token = await getToken();
+  const controller = new AbortController();
+  const timeoutID = setTimeout(() => controller.abort(), timeoutMs);
+  try {
+    return await fetch(input, {
+      ...init,
+      signal: mergeSignals([init.signal, controller.signal]),
+      headers: {
+        ...(init.headers ?? {}),
+        Authorization: `Bearer ${token}`,
+      },
+    });
+  } finally {
+    clearTimeout(timeoutID);
+  }
+}
